🖊 Completed basic UI.
